# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ante",
    "description": "AI-Native Trading Engine API",
    "version": "0.9.0"
  },
  "paths": {
...
    "/api/data/datasets": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "List Datasets",
        "description": "보유 데이터셋 목록 (페이지네이션 지원).\n\ndata_type 미지정 시 모든 유형을 반환하고, 지정 시 해당 유형만 반환한다.\n\n종목 목록만 먼저 수집하여 페이지네이션을 적용한 뒤,\n해당 페이지의 종목에 대해서만 date_range를 계산한다.\nrow_count와 file_size는 상세 조회 전용이므로 기본값(0)을 반환한다.\n\nReturns:\n    {items: [...], total: int}",
        "operationId": "list_datasets_api_data_datasets_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Timeframe"
            }
          },
          {
            "name": "data_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ohlcv",
                    "fundamental"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "데이터 유형 (ohlcv, fundamental, 미지정 시 전체)",
              "title": "Data Type"
            },
            "description": "데이터 유형 (ohlcv, fundamental, 미지정 시 전체)"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/datasets/{dataset_id}": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Get Dataset Detail",
        "description": "데이터셋 상세 조회 (메타데이터 + 최근 5행 미리보기). 인증된 master/human\n또는 ``data:read`` scope 를 보유한 agent 만 호출 가능 (#1407).\n\ndataset_id 형식: \"{symbol}__{timeframe}\" (예: \"005930__1d\")",
        "operationId": "get_dataset_detail_api_data_datasets__dataset_id__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "data"
        ],
        "summary": "Delete Dataset",
        "description": "데이터셋 삭제. 인증된 master/human 또는 ``data:write`` scope 를 보유한\nagent 만 호출 가능 (#1407).\n\ndataset_id 형식: \"{symbol}__{timeframe}\" (예: \"005930__1d\")\nfundamental의 경우: \"{symbol}__fundamental\" (예: \"005930__fundamental\")",
        "operationId": "delete_dataset_api_data_datasets__dataset_id__delete",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "data_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "데이터 유형 (ohlcv, fundamental)",
              "default": "ohlcv",
              "title": "Data Type"
            },
            "description": "데이터 유형 (ohlcv, fundamental)"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid dataset_id format",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Dataset not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Data store not available",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/schema": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Get Data Schema",
        "description": "데이터 스키마 조회. data_type에 따라 해당 스키마를 반환. 인증된\nmaster/human 또는 ``data:read`` scope 를 보유한 agent 만 호출 가능\n(#1407).",
        "operationId": "get_data_schema_api_data_schema_get",
        "parameters": [
          {
            "name": "data_type",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "ohlcv",
                "fundamental"
              ],
              "type": "string",
              "description": "데이터 유형 (ohlcv, fundamental)",
              "default": "ohlcv",
              "title": "Data Type"
            },
            "description": "데이터 유형 (ohlcv, fundamental)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSchemaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataSchemaResponse": {
        "properties": {},
        "additionalProperties": true,
        "type": "object",
        "title": "DataSchemaResponse",
        "description": "데이터 스키마 응답 (동적 키-값)."
      },
      "DatasetDetailResponse": {
        "properties": {
          "dataset": {
            "$ref": "#/components/schemas/DatasetItem"
          },
          "preview": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Preview",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "dataset"
        ],
        "title": "DatasetDetailResponse",
        "description": "데이터셋 상세 응답 (메타데이터 + 미리보기)."
      },
      "DatasetItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe"
          },
          "data_type": {
            "type": "string",
            "title": "Data Type"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count",
            "default": 0
          },
          "file_size": {
            "type": "integer",
            "title": "File Size",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "symbol",
          "timeframe",
          "data_type"
        ],
        "title": "DatasetItem",
        "description": "데이터셋 아이템."
      },
      "DatasetListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DatasetItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "DatasetListResponse",
        "description": "데이터셋 목록 응답."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "ErrorResponse": {
        "description": "RFC 7807 Problem Details 에러 응답.",
        "properties": {
          "type": {
            "default": "/errors/internal",
            "title": "Type",
            "type": "string"
          },
          "title": {
            "default": "Internal Server Error",
            "title": "Title",
            "type": "string"
          },
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          },
          "status": {
            "default": 500,
            "title": "Status",
            "type": "integer"
          },
          "instance": {
            "default": "",
            "title": "Instance",
            "type": "string"
          }
        },
        "title": "ErrorResponse",
        "type": "object"
      },
...
    }
  }
}